requirements.txt
asgiref==3.6.0
asttokens==2.1.0
autopep8==2.0.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
debugpy==1.6.5
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
ipykernel==6.20.1
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
packaging==23.0
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.31
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.7
pywin32==305
pyzmq==24.0.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.3
sqlparse==0.4.3
stack-data==0.6.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.17.1
wcwidth==0.2.5
Werkzeug==2.2.2
